requirements.txt
altgraph==0.16.1
cffi==1.11.5
future==0.16.0
macholib==1.11
numpy==1.18.4
pefile==2018.8.8
Pillow==7.1.2
pycparser==2.19
PyInstaller==3.4
Pypubsub==4.0.0
PySoundFile==0.9.0.post1
six==1.11.0
wxPython==4.1.0
